# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "backend",
    "description": "12-factor Rust backend: layered store + RBAC + JWT + WS + worker",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/pickup-points": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "`GET /api/admin/pickup-points?routeId=` — list pickup points for a route.",
        "operationId": "list",
        "parameters": [
          {
            "name": "routeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pickup point list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPickupPointListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "`POST /api/admin/pickup-points` — create a pickup point.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPickupPointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMutationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/admin/pickup-points/{id}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "`PUT /api/admin/pickup-points/{id}` — update a pickup point.",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Pickup point ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPickupPointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMutationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "`DELETE /api/admin/pickup-points/{id}` — delete a pickup point.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Pickup point ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/admin/reviews": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "`GET /api/admin/reviews` — list reviews with admin filters.",
        "operationId": "list",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "brandId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "routeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/admin/reviews/{id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "`DELETE /api/admin/reviews/{id}` — delete a review (admin override).",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Review ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "`PATCH /api/admin/reviews/{id}` — moderate a review (status + reply).",
        "operationId": "moderate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Review ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Moderated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerateReviewResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/reviews": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "`GET /api/reviews` — list reviews with optional filters.",
        "operationId": "list",
        "parameters": [
          {
            "name": "brand_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "route_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "`POST /api/reviews` — create a review. Requires authentication.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewMutationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/reviews/{id}": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "`GET /api/reviews/{id}` — get a review by ID.",
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Review ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "reviews"
        ],
        "summary": "`DELETE /api/reviews/{id}` — delete a review. Requires authentication.",
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Review ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDeleteResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "patch": {
        "tags": [
          "reviews"
        ],
        "summary": "`PATCH /api/reviews/{id}` — update a review. Requires authentication.",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Review ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewMutationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminMutationResponse": {
        "type": "object",
        "description": "Response of `POST /api/admin/brands` + `PUT /api/admin/brands/{id}` + `DELETE`.",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "AdminPickupPointListResponse": {
        "type": "object",
        "description": "Response of `GET /api/admin/pickup-points`.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminPickupPointOut"
            }
          }
        }
      },
      "AdminPickupPointOut": {
        "type": "object",
        "required": [
          "id",
          "routeId",
          "stopOrder",
          "createdAt"
        ],
        "properties": {
          "address": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "lat": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "lon": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "routeId": {
            "type": "string",
            "format": "uuid"
          },
          "stopOrder": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "AdminReviewListResponse": {
        "type": "object",
        "description": "Response of `GET /api/admin/reviews`.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewOut"
            }
          }
        }
      },
...
      "CreateReviewInput": {
        "type": "object",
        "description": "Input for creating a review. Field names on the wire are camelCase\n(so the frontend can send `{ bookingId, routeId, brandId, rating, ... }`)\nbut the Rust struct uses snake_case.",
        "required": [
          "rating"
        ],
        "properties": {
          "authorName": {
            "type": [
              "string",
              "null"
            ]
          },
          "authorPhone": {
            "type": [
              "string",
              "null"
            ]
          },
          "bookingId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "brandId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "photos": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "routeId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "tripSessionId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "userId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Overwritten by the server from the authenticated user."
          }
        }
      },
...
      "ModerateReviewRequest": {
        "type": "object",
        "description": "Request body for `PATCH /api/admin/reviews/{id}` (moderation).",
        "properties": {
          "brandReply": {
            "type": [
              "string",
              "null"
            ],
            "description": "Brand's reply text (sets `replied_at` automatically)."
          },
          "status": {
            "type": [
              "string",
              "null"
            ],
            "description": "`pending` | `approved` | `rejected` | `hidden`"
          }
        }
      },
      "ModerateReviewResponse": {
        "type": "object",
        "description": "Response of `PATCH /api/admin/reviews/{id}`.",
        "required": [
          "id",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "ReviewDeleteResponse": {
        "type": "object",
        "description": "Response of `DELETE /api/reviews/{id}`.",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
      "ReviewListResponse": {
        "type": "object",
        "description": "Response of `GET /api/reviews`.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewOut"
            }
          }
        }
      },
      "ReviewMutationResponse": {
        "type": "object",
        "description": "Response of `POST /api/reviews` and `PATCH /api/reviews/{id}`.",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ReviewOut": {
        "type": "object",
        "description": "A review row, as returned by `GET /api/reviews` and `GET /api/reviews/{id}`.",
        "required": [
          "id",
          "rating",
          "status",
          "helpfulCount",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "authorName": {
            "type": [
              "string",
              "null"
            ]
          },
          "authorPhone": {
            "type": [
              "string",
              "null"
            ]
          },
          "bookingId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "brandId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "helpfulCount": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "photos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rating": {
            "type": "integer",
            "format": "int64",
            "description": "1..=5"
          },
          "repliedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "reply": {
            "type": [
              "string",
              "null"
            ],
            "description": "Brand's reply (admin-written)."
          },
          "routeId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "description": "`pending` | `approved` | `rejected`"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "tripSessionId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "updatedAt": {
            "type": "string"
          },
          "userId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
...
      "UpdateReviewInput": {
        "type": "object",
        "description": "Input for updating a review.",
        "properties": {
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "photos": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "rating": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UpsertPickupPointRequest": {
        "type": "object",
        "description": "Request body for `POST /api/admin/pickup-points` + `PUT /api/admin/pickup-points/{id}`.",
        "properties": {
          "address": {
            "type": [
              "string",
              "null"
            ]
          },
          "kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "lat": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "lon": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "routeId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "stopOrder": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
    }
  },
...
}